+++ viewer/__mocks__/html2canvas-pro.js
@@ -1,0 +1,8 @@
+// Mock for html2canvas-pro module
+export default jest.fn(() => Promise.resolve({
+  toBlob: jest.fn((callback) => {
+    // Mock blob creation
+    const blob = new Blob(['mock-image'], { type: 'image/png' });
+    callback(blob);
+  })
+}));
